Move to utils file
